requirements.txt
asgiref==3.8.1
asttokens==3.0.0
certifi==2025.4.26
charset-normalizer==3.4.3
colorama==0.4.6
comm==0.2.2
debugpy==1.8.11
decorator==5.1.1
distlib==0.3.9
dj-database-url==3.0.1
Django==5.0.7
django-cors-headers==4.7.0
django-environ==0.11.2
django-extensions==4.1
django-mathfilters==1.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
executing==2.1.0
filelock==3.18.0
gunicorn==22.0.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
parso==0.8.4
passlib==1.7.4
pillow==10.4.0
pipenv==2025.0.2
platformdirs==4.3.8
prompt_toolkit==3.0.48
psutil==6.1.1
psycopg2-binary==2.9.10
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.9.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
pywin32==308
pyzmq==26.2.0
qrcode==7.4.2
requests==2.32.3
setuptools==80.4.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.31.2
wcwidth==0.2.13
whitenoise==6.7.0
